contact structure created, mentor fix
